Accept doLamps option to expose command, control dcb if set.
The `doLamps` option signals that we should tell the dcb (or pfilamps)
actor to fire up the lamps after the shutter has been opened. Exposure
timing is then set by the lamps, not the shutter.

For now, the shutter command blocks for the full exposure time. So we
tell the dcb lamps to fire a couple of seconds after we tell them to,
then open the shutter. This will be fixed.